# src/frontend/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ocelescope Backend",
    "version": "0.1"
  },
  "paths": {
...
    "/tasks/system": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "returns all tasks of a session",
        "operationId": "getSystemTasks",
        "parameters": [
          {
            "name": "task_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Name"
            }
          },
          {
            "name": "task_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "default": [],
              "title": "Task Ids"
            }
          },
          {
            "name": "only_running",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Only Running"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SystemTaskSummary"
                  },
                  "title": "Response Getsystemtasks"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/system/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "returns the task of a given taskId",
        "operationId": "getSystemTask",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemTaskSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/plugin": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "returns all tasks of a session",
        "operationId": "getPluginTasks",
        "parameters": [
          {
            "name": "plugin_id",
            "in": "query",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plugin Id"
            }
          },
          {
            "name": "method_name",
            "in": "query",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Method Name"
            }
          },
          {
            "name": "only_running",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Only Running"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PluginTaskSummary"
                  },
                  "title": "Response Getplugintasks"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/plugin/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "returns the task of a given taskId",
        "operationId": "getPluginTask",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginTaskSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/resources/": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Get Resources",
        "operationId": "resources",
        "parameters": [
          {
            "name": "resource_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResourceApi"
                  },
                  "title": "Response Resources"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "resources"
        ],
        "summary": "Upload Resource",
        "operationId": "uploadResource",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadResource"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/resources/resource/{resource_id}/download": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Download Output",
        "operationId": "download_output_resources_resource__resource_id__download_get",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/resources/resource/{resource_id}": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Get Resource",
        "operationId": "resource",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetResourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "resources"
        ],
        "summary": "Delete Resource",
        "operationId": "deleteResource",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "resources"
        ],
        "summary": "Rename Resource",
        "operationId": "renameResource",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          },
          {
            "name": "new_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "New Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_uploadResource": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_uploadResource"
      },
...
      "DotVis": {
        "properties": {
          "type": {
            "type": "string",
            "const": "dot",
            "title": "Type"
          },
          "dot_str": {
            "type": "string",
            "title": "Dot Str"
          },
          "layout_engine": {
            "type": "string",
            "enum": [
              "circo",
              "dot",
              "fdp",
              "sfdp",
              "neato",
              "osage",
              "patchwork",
              "twopi",
              "nop",
              "nop2"
            ],
            "title": "Layout Engine"
          }
        },
        "type": "object",
        "required": [
          "type",
          "dot_str",
          "layout_engine"
        ],
        "title": "DotVis"
      },
...
      "GetResourceResponse": {
        "properties": {
          "resource": {
            "$ref": "#/components/schemas/ResourceApi"
          },
          "visualization": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/Graph"
                  },
                  {
                    "$ref": "#/components/schemas/Table"
                  },
                  {
                    "$ref": "#/components/schemas/SVGVis"
                  },
                  {
                    "$ref": "#/components/schemas/DotVis"
                  }
                ],
                "discriminator": {
                  "propertyName": "type",
                  "mapping": {
                    "dot": "#/components/schemas/DotVis",
                    "graph": "#/components/schemas/Graph",
                    "svg": "#/components/schemas/SVGVis",
                    "table": "#/components/schemas/Table"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Visualization"
          }
        },
        "type": "object",
        "required": [
          "resource",
          "visualization"
        ],
        "title": "GetResourceResponse"
      },
      "Graph": {
        "properties": {
          "type": {
            "type": "string",
            "const": "graph",
            "title": "Type"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "type": "array",
            "title": "Edges"
          },
          "layout_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphvizLayoutConfig"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "type",
          "nodes",
          "edges"
        ],
        "title": "Graph"
      },
      "GraphEdge": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "arrows": {
            "prefixItems": [
              {
                "anyOf": [
                  {
                    "type": "string",
                    "enum": [
                      "triangle",
                      "circle-triangle",
                      "triangle-backcurve",
                      "tee",
                      "circle",
                      "chevron",
                      "triangle-tee",
                      "triangle-cross",
                      "vee",
                      "square",
                      "diamond"
                    ]
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              {
                "anyOf": [
                  {
                    "type": "string",
                    "enum": [
                      "triangle",
                      "circle-triangle",
                      "triangle-backcurve",
                      "tee",
                      "circle",
                      "chevron",
                      "triangle-tee",
                      "triangle-cross",
                      "vee",
                      "square",
                      "diamond"
                    ]
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Arrows"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "source",
          "target",
          "arrows"
        ],
        "title": "GraphEdge"
      },
      "GraphNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "shape": {
            "type": "string",
            "enum": [
              "circle",
              "triangle",
              "rectangle",
              "diamond",
              "hexagon"
            ],
            "title": "Shape"
          },
          "width": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "x": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "X"
          },
          "y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y"
          },
          "border_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Border Color"
          },
          "label_pos": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "top",
                  "center",
                  "bottom"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Label Pos"
          }
        },
        "type": "object",
        "required": [
          "id",
          "shape"
        ],
        "title": "GraphNode",
        "description": "[TODO:description]\n\nAttributes:\n    id: [TODO:attribute]\n    label: [TODO:attribute]\n    shape: [TODO:attribute]\n    width: [TODO:attribute]\n    height: [TODO:attribute]\n    color: [TODO:attribute]\n    x: [TODO:attribute]\n    y: [TODO:attribute]\n    border_color: [TODO:attribute]\n    label_pos: [TODO:attribute]"
      },
      "GraphvizLayoutConfig": {
        "properties": {
          "engine": {
            "type": "string",
            "enum": [
              "circo",
              "dot",
              "fdp",
              "sfdp",
              "neato",
              "osage",
              "patchwork",
              "twopi",
              "nop",
              "nop2"
            ],
            "title": "Engine",
            "default": "dot"
          },
          "graphAttrs": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "boolean"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graphattrs"
          },
          "nodeAttrs": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "boolean"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nodeattrs"
          },
          "edgeAttrs": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "boolean"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edgeattrs"
          }
        },
        "type": "object",
        "title": "GraphvizLayoutConfig"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PluginOutput": {
        "properties": {
          "ocel_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ocel Ids",
            "default": []
          },
          "resource_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Resource Ids",
            "default": []
          }
        },
        "type": "object",
        "title": "PluginOutput"
      },
      "PluginSource": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "plugin_name": {
            "type": "string",
            "title": "Plugin Name"
          },
          "method_name": {
            "type": "string",
            "title": "Method Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "plugin_name",
          "method_name",
          "version"
        ],
        "title": "PluginSource"
      },
      "PluginTaskSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "state": {
            "$ref": "#/components/schemas/TaskState"
          },
          "plugin_id": {
            "type": "string",
            "title": "Plugin Id"
          },
          "method_name": {
            "type": "string",
            "title": "Method Name"
          },
          "output": {
            "$ref": "#/components/schemas/PluginOutput"
          }
        },
        "type": "object",
        "required": [
          "id",
          "state",
          "plugin_id",
          "method_name",
          "output"
        ],
        "title": "PluginTaskSummary"
      },
...
      "ResourceApi": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PluginSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "type",
          "name",
          "source",
          "id"
        ],
        "title": "ResourceApi"
      },
...
      "SVGVis": {
        "properties": {
          "type": {
            "type": "string",
            "const": "svg",
            "title": "Type"
          },
          "svg": {
            "type": "string",
            "title": "Svg"
          }
        },
        "type": "object",
        "required": [
          "type",
          "svg"
        ],
        "title": "SVGVis"
      },
      "SystemTaskSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "state": {
            "$ref": "#/components/schemas/TaskState"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "state",
          "name",
          "metadata"
        ],
        "title": "SystemTaskSummary"
      },
      "Table": {
        "properties": {
          "type": {
            "type": "string",
            "const": "table",
            "title": "Type"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/TableColumn"
            },
            "type": "array",
            "title": "Columns"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "type",
          "columns",
          "rows"
        ],
        "title": "Table"
      },
...
      "TaskState": {
        "type": "string",
        "enum": [
          "PENDING",
          "STARTED",
          "SUCCESS",
          "FAILURE",
          "SKIPPED",
          "CANCELLED"
        ],
        "title": "TaskState"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}